cmdFiles/requirements.txt
biopython==1.76
mysql-connector==2.2.9
numpy==1.18.2
pkg-resources==0.0.0
